Modifications to get less warnings.

+++ src/prototypes.h
@@ -117,7 +117,7 @@ void space_text(void);
 void space_text_balance_nested_parens(void);
 int space_col_align(chunk_t *first, chunk_t *second);
 int space_needed(chunk_t *first, chunk_t *second);
-void space_add_after(chunk_t *pc, int count);
+void space_add_after(chunk_t *pc, size_t count);
 
 
 /*
